requirements.txt
python-telegram-bot==20.8
python-dotenv==1.0.1
opencv-python-headless==4.10.0.84
numpy==2.3.4
